% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_scProfiles.R
\name{aggregate_scProfiles}
\alias{aggregate_scProfiles}
\title{Aggregate scRNA Profiles}
\usage{
aggregate_scProfiles(scRNA_exp, scRNA_st, cell_id_col, group_id_col)
}
\arguments{
\item{scRNA_exp}{scRNA expression profiles}

\item{scRNA_st}{scRNA expression profiles sample table}

\item{cell_id_col}{column name for cell id}

\item{group_id_col}{column name for cluster label or cell type label}
}
\value{
a list consists of aggregated expression profiles and aggregated sample table
}
\description{
Aggregate scRNA expression profiles that have the same cluster label or cell type label
}
