% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgGeneCat.R
\name{avgGeneCat}
\alias{avgGeneCat}
\title{Average Category Gene Expressions}
\usage{
avgGeneCat(expDat, sampTab, dLevel, sampID)
}
\arguments{
\item{expDat}{the normalized expression table from \code{\link{trans_prop}}}

\item{sampTab}{the sample table}

\item{dLevel}{the name of the column that contains categories}

\item{sampID}{the name of the column that contains sample ID}
}
\value{
a matrix containing the average of
}
\description{
Averaging the gene expression level for each category
}
