% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlotGene.R
\name{barPlotGene}
\alias{barPlotGene}
\title{Plot Specific Gene Expression}
\usage{
barPlotGene(
  GeneCompareTab,
  sdGeneCat = NULL,
  gene,
  text_size = 10,
  title_size = 12
)
}
\arguments{
\item{GeneCompareTab}{the gene comparsion matrix generated from \code{\link{makeGeneCompareTab}}}

\item{sdGeneCat}{the standard deviation matrix generated from \code{\link{sdGeneCat}}}

\item{gene}{the name of the specific gene of interest}

\item{text_size}{the size of the ticks}

\item{title_size}{the size of the labels}
}
\value{
expression bar plot
}
\description{
Plots a specific gene and its expression levels across the sampels
}
