% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadClass_ensemblePredict.R
\name{broadClass_ensemblePredict}
\alias{broadClass_ensemblePredict}
\title{Predict query using an ensemble of broad class classifier}
\usage{
broadClass_ensemblePredict(broadReturnList, queryData, nrand = 2)
}
\arguments{
\item{broadReturnList}{a list of broadReturn from training}

\item{queryData}{the expression data of query data}

\item{nrand}{the number of random profiles generate for evaluation process}
}
\value{
an average classification score matrix
}
\description{
The function predicts the query data using an ensemble broad class classifier
}
