% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadClass_ensembleTrain.R
\name{broadClass_ensembleTrain}
\alias{broadClass_ensembleTrain}
\title{Broad Class Ensemble Training (not commonly used)}
\usage{
broadClass_ensembleTrain(
  stTrain,
  expTrain,
  colName_cat,
  colName_samp = "row.names",
  numClassifier = 5,
  ncells = 60,
  nTopGenes = 20,
  nTopGenePairs = 50,
  nRand = 40,
  nTrees = 1000,
  weightedDown_total = 5e+05,
  weightedDown_dThresh = 0.25,
  transprop_xFact = 1e+05
)
}
\arguments{
\item{stTrain}{a dataframe that matches the samples with category}

\item{expTrain}{the expression matrix}

\item{colName_cat}{the name of the column that contains categories}

\item{colName_samp}{the name of the column that contains sample names}

\item{numClassifier}{the number of classifiers in the ensemble}

\item{ncells}{the number of samples per category for training}

\item{nTopGenes}{the number of classification genes per category}

\item{nTopGenePairs}{the number of top gene pairs per category}

\item{nRand}{number of random profiles generate for training}

\item{nTrees}{number of trees for random forest classifier}

\item{weightedDown_dThresh}{the threshold at which anything lower than that is 0 for weighted_down function}

\item{transprop_xFact}{scaling factor for transprop}

\item{weightDown_total}{numeric post transformation sum of read counts for weighted_down function}
}
\value{
a list of broadclass_returns with classifiers
}
\description{
Tranining an ensemble of broad class classifiers
}
