% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadClass_predict.R
\name{broadClass_predict}
\alias{broadClass_predict}
\title{Predict query using broad class classifier}
\usage{
broadClass_predict(cnProc, expDat, nrand = 2)
}
\arguments{
\item{cnProc}{the cnProc of the broad classifier}

\item{expDat}{the expression data of query data}

\item{nrand}{the number of random profiles generate for evaluation process}
}
\value{
a classification score matrix
}
\description{
The function predicts the query data using the broad class classifier
}
