% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadClass_train.R
\name{broadClass_train}
\alias{broadClass_train}
\title{Broad Class Training}
\usage{
broadClass_train(
  stTrain,
  expTrain,
  colName_cat,
  colName_samp = "row.names",
  nTopGenes = 20,
  nTopGenePairs = 50,
  nRand = 40,
  nTrees = 1000,
  stratify = FALSE,
  sampsize = 40,
  weightedDown_total = 5e+05,
  weightedDown_dThresh = 0.25,
  transprop_xFact = 1e+05,
  quickPairs = FALSE,
  coreProportion = 1/4
)
}
\arguments{
\item{stTrain}{a dataframe that matches the samples with category}

\item{expTrain}{the expression matrix}

\item{colName_cat}{the name of the column that contains categories}

\item{colName_samp}{the name of the column that contains sample names}

\item{nTopGenes}{the number of classification genes per category}

\item{nTopGenePairs}{the number of top gene pairs per category}

\item{nRand}{number of random profiles generate for training}

\item{nTrees}{number of trees for random forest classifier}

\item{stratify}{TRUE if stratified sampling}

\item{sampsize}{sample size for stratified sampling}

\item{weightedDown_dThresh}{the threshold at which anything lower than that is 0 for weighted_down function}

\item{transprop_xFact}{scaling factor for transprop}

\item{quickPairs}{TRUE if using quick version of genepair transform}

\item{coreProportion}{the proportion of logical cores for finding classification genes and top scoring gene pairs. If you want to disable parallel processing, then enter 0}

\item{weightDown_total}{numeric post transformation sum of read counts for weighted_down function}
}
\value{
a list containing normalized expression data, classification gene list, cnPRoc
}
\description{
Tranining broad class classifier
}
