% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_tfScores.R
\name{calc_tfScores}
\alias{calc_tfScores}
\title{Calculate the TF
Calculate the TF score given all the parameters}
\usage{
calc_tfScores(tf, targs, sampleID, z_scoreMat, netGenes, weights, grnTable,
  correlationFactor = 1)
}
\arguments{
\item{tf}{TF}

\item{targs}{a list of targets for TF}

\item{sampleID}{the sample ID}

\item{z_scoreMat}{the z value matrix}

\item{netGenes}{genes selected for building the subnetwork}

\item{weights}{a vector of weights for each gene}

\item{grnTable}{the grn table with correlation values and signs}

\item{correlationFactor}{the weight of correlation between}
}
\description{
Calculate the TF
Calculate the TF score given all the parameters
}
