% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn_classAssess.R
\name{ccn_classAssess}
\alias{ccn_classAssess}
\title{Assessment of Classifier}
\usage{
ccn_classAssess(ct_scores, stVal, classLevels = "description2",
  dLevelSID = "sample_id", resolution = 0.005)
}
\arguments{
\item{ct_scores}{matrix of classification scores}

\item{stVal}{sample table}

\item{classLevels}{column name of the sample table indicating the categories to be used as ground truth to assess classifiers}

\item{dLevelSID}{column name of the sample table that indicates sample id}

\item{resolution}{increment of thresholds for assessing the classifier}
}
\value{
list of data frames with threshold, sens, precision
}
\description{
Assess classifiers using validation data
}
