% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_construction.R
\name{ccn_extractRegsDF}
\alias{ccn_extractRegsDF}
\title{extracts the TRs, zscores, and corr values passing thresh}
\usage{
ccn_extractRegsDF(zscores, corrMatrix, genes, threshold)
}
\arguments{
\item{zscores}{zscore matrix, non-TFs already removed from columns}

\item{corrMatrix}{correlation matrix}

\item{genes}{vector of target genes}

\item{threshold}{zscore threshold}
}
\value{
data.frame(target=targets, reg=regulators, zscore=zscoresX, corr=correlations);
}
\description{
extracts the TRs, zscores, and corr values passing thresh
}
