% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_construction.R
\name{ccn_get_targets}
\alias{ccn_get_targets}
\title{get targets of a tf}
\usage{
ccn_get_targets(aGraph)
}
\arguments{
\item{aGraph}{an iGraph}
}
\value{
target list
}
\description{
get targets of a tf
}
