% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_construction.R
\name{ccn_makeGRN}
\alias{ccn_makeGRN}
\title{make GRN}
\usage{
ccn_makeGRN(
  expTrain,
  stTrain,
  dLevel,
  zThresh = 4,
  dLevelGK = NULL,
  prune = FALSE,
  holm = 1e-04,
  cval = 0.4
)
}
\arguments{
\item{expTrain}{normalized expression matrix}

\item{stTrain}{sample table}

\item{dLevel}{name of the column with different categories}

\item{zThresh}{threshold of z score for CLR network reconstruction}

\item{dLevelGK}{name of the column with different germlayer categories}

\item{prune}{boolean limit to genes exclusively detected as CT in one CT}

\item{holm}{threshold of holm adjusted p value for selecting subnetwork genes}

\item{cval}{threshold of cval for selectin subnetwork genes. Higher cval indicates selecting higher enriched genes}
}
\value{
constructed GRN
}
\description{
Construct GRN
}
