% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_construction.R
\name{ccn_MakeTLs}
\alias{ccn_MakeTLs}
\title{get targets of tFs}
\usage{
ccn_MakeTLs(graphList)
}
\arguments{
\item{graphList}{a list of networks represented as iGraphs}
}
\value{
list of tf=>targets
}
\description{
get targets of tFs
}
