% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_status.R
\name{ccn_queryGRNstatus}
\alias{ccn_queryGRNstatus}
\title{Get Query GRN status}
\usage{
ccn_queryGRNstatus(
  expQuery,
  expTrain,
  stTrain,
  dLevel,
  sidCol,
  grn_return,
  trainNorm = NULL,
  classifier_return,
  classWeight = TRUE,
  exprWeight = FALSE,
  prune = TRUE,
  xmax = 1000,
  predSD = FALSE
)
}
\arguments{
\item{expQuery}{logRanked query expression matrix}

\item{expTrain}{logRanked training expression matrix}

\item{stTrain}{sample table of training expression matrix}

\item{dLevel}{the name of the column with cancer types}

\item{sidCol}{the name of the column with sample IDs}

\item{grn_return}{the grn list that is returned from \code{\link{ccn_makeGRN}}}

\item{trainNorm}{normalization statistics from \code{\link{ccn_trainNorm}}. If you are using pre-calculated normalization statistics please make sure all the parameters are the same for applying to query and calculating normalization}

\item{classifier_return}{the classifier_return list that is returned from \code{\link{broadClass_train}}}

\item{classWeight}{boolean indicating whether to take the importance of the classification into status calculation}

\item{exprWeight}{boolean indicating whether to take the weight of gene expression into status calculation}

\item{prune}{boolean indicating whether to select exclusive genes for processing classification gene importance}

\item{predSD}{a parameter for calculating normalization statistics from training data}
}
\value{
a matrix indicating the GRN status
}
\description{
Get the GRN status of query samples
}
