% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_status.R
\name{ccn_reduceMatLarge}
\alias{ccn_reduceMatLarge}
\title{reduce large matrix
reduce large matrix from ccn_extract_SN_DF}
\usage{
ccn_reduceMatLarge(
  datFrame,
  valCol = "score",
  cName = "description",
  iterOver = "subNet"
)
}
\arguments{
\item{datFrame}{the result from ccn_extract_SN_DF}

\item{valCol}{the column name of the score}

\item{cName}{the column name of the different groups}

\item{iterOver}{the column name of the subnetworks}
}
\value{
reduced large matrix
}
\description{
reduce large matrix
reduce large matrix from ccn_extract_SN_DF
}
