% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_construction.R
\name{ccn_sampR_to_pattern}
\alias{ccn_sampR_to_pattern}
\title{return a pattern for use in ccn_testPattern (template matching)}
\usage{
ccn_sampR_to_pattern(sampR)
}
\arguments{
\item{sampR}{vector}
}
\value{
ans
}
\description{
return a pattern for use in ccn_testPattern (template matching)
}
