% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_construction.R
\name{ccn_specGRNs}
\alias{ccn_specGRNs}
\title{extract sub-networks made up of CT genes;}
\usage{
ccn_specGRNs(rawGRNs, specGenes)
}
\arguments{
\item{rawGRNs}{result of running ccn_getRawGRN}

\item{specGenes}{result of running ccn_specGenesAll}
}
\value{
list(geneLists=geneLists, graphLists=graphLists, tfTargets=tfTargets)
}
\description{
extract sub-networks made up of CT genes;
}
