% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_tfScores.R
\name{ccn_tfScores}
\alias{ccn_tfScores}
\title{calculate TF scores}
\usage{
ccn_tfScores(expQuery, subnetName, grnAll, trainNorm, geneImportance,
  classWeight = TRUE, classWeightVal = 3, exprWeight = TRUE,
  exprWeightVal = 3, correlationFactor = 1)
}
\arguments{
\item{expQuery}{the logRanked expression matrix for query data}

\item{subnetName}{the name of the target cell type}

\item{grnAll}{result of running \code{\link{ccn_makeGRN}}}

\item{trainNorm}{result of running \code{\link{ccn_trainNorm}}}

\item{geneImportance}{the importance of gene pairs from classifier. Result of running \code{\link{processImportance}}}

\item{classWeight}{whether to take the weight of the classifier into calculation}

\item{classWeightVal}{the value of the classifier weight}

\item{exprWeight}{whether to take the weight of the expression into calculation}

\item{exprWeightVal}{the value of the expression weight}

\item{correlationFactor}{the weight of correlation direction in the network}
}
\value{
matrix of TF scores and query samples
}
\description{
calculate the TF that one should target to yield target cell type
}
