% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_tfScores.R
\name{ccn_zscoreVect}
\alias{ccn_zscoreVect}
\title{Compute mean Z scores of given genes}
\usage{
ccn_zscoreVect(genes, xvals, tVals, ctt)
}
\arguments{
\item{genes}{the genes}

\item{xvals}{named vector}

\item{tVals}{mean and SD of genes in training data}

\item{ctt}{cell type}
}
\value{
a vector of Z scores
}
\description{
Compute the mean Z score of given genes in each sample
}
