<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn_classAssess.R
\name{cn_eval}
\alias{cn_eval}
\title{determine performance of classification at given threshold}
\usage{
cn_eval(vect, sampTab, dLevel, classification, threshs = seq(0, 1, by =
  0.05), dLevelSID = "sample_id")
}
\arguments{
\item{vect}{vector of values}

\item{sampTab}{sample table}

\item{dLevel}{colname}

\item{classification}{actual classification}

\item{dLevelSID}{column to indicate sample id}

\item{thresh}{threshold above which to make a call}
}
\value{
vector of TP FN FP TN
}
\description{
determine performance of classification at given threshold
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{cn_eval}
\alias{cn_eval}
\title{run cn_clPerf across thresholds}
\usage{
cn_eval(vect, sampTab, dLevel, classification, threshs = seq(0, 1, by = 0.05),
  dLevelSID = "sample_id")
}
\arguments{
\item{vect}{named vector}

\item{dLevel}{description level}

\item{classification}{classification matrix}

\item{threshs}{seq of pval cutoffs}

\item{dLevelSID}{column to indicate sample id}
}
\value{
return a data frame of the number of TP, FN, FP, and TN, and pval cutoff
}
\description{
run cn_clPerf across thresholds
}
>>>>>>> parent of 590b901... Documentation
