% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{cn_HmClass}
\alias{cn_HmClass}
\title{simple heatmap of the classification result}
\usage{
cn_HmClass(classRes, isBig = FALSE)
}
\arguments{
\item{isBig}{is this a big heatmap? TRUE or FALSE}

\item{cnRes}{returned from cn_sapply}
}
\value{
nothing
}
\description{
Heatmap of the classification result.
}
\examples{
cn_HmClass(cnRes, isBig=TRUE)

}
