% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversionList.R
\docType{data}
\name{conversionList}
\alias{conversionList}
\title{List of Gene Conversion dataframe}
\format{A list with 3 dataframes, which are:
\describe{
\item{ensTransToGene}{Dataframe that maps Ensembl Transcript ID to Gene Symbol}
\item{ensGeneToGene}{Dataframe that maps Ensembl Gene ID to Gene Symbol}
\item{mouseGeneToGene}{Dataframe that maps Mouse Gene Symbol to Human Gene Symbol}
}}
\usage{
conversionList
}
\description{
A list containing 3 dataframes that maps Ensembl Gene ID, Ensembl Transcript ID and Mouse Gene Symbols to Human Gene Symbols
}
\keyword{datasets}
