% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_construction.R
\name{find_tfs}
\alias{find_tfs}
\title{find transcript factors}
\usage{
find_tfs(species = "Hs")
}
\arguments{
\item{species}{defaul is 'Hs', can also be 'Mm;}
}
\value{
vector fo TF names
}
\description{
find transcript factors
}
