% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptGetTop.R
\name{findBestPairs}
\alias{findBestPairs}
\title{Find best pairs}
\usage{
findBestPairs(xdiff, n = 50, maxPer = 3)
}
\arguments{
\item{xdiff}{statList of a certain group generated as an intermediate step from \code{\link{ptGetTop}}}

\item{n}{the number of top pairs}

\item{maxPer}{indicates the maximum number of pairs that a gene is allowed to be in}
}
\value{
vector of suitable gene pairs
}
\description{
Perform finding the best and diverse set of gene pairs for training
}
