% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subClass_hmClass.R
\name{findCategory}
\alias{findCategory}
\title{Find the broad category}
\usage{
findCategory(ThisColumn, thresHold_list)
}
\arguments{
\item{ThisColumn}{the individual sample in a matrix}

\item{thresHold_list}{the list of threshold for each category}
}
\value{
classification heatmap
}
\description{
Find the broad cateogry based on a threshold
}
