% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn_analysis.R
\name{gene_expr_sum}
\alias{gene_expr_sum}
\title{Sum transcription expression estimates to gene-level expression measures}
\usage{
gene_expr_sum(expRaw, numCores = 3,
  geneTabfname = "./geneToTrans_Mus_musculus.GRCm38.80.exo_Jun_02_2015.R",
  nameCol = "gene_name")
}
\arguments{
\item{expRaw}{salmon isoform estimates}

\item{numCores}{num of cores to use for parallel}

\item{geneTabfname}{gene <-> transcript mapping, df that needs cols: gene_id, transcript_id}

\item{nameCol}{gene ann table column name to average over}
}
\value{
matrix (ansCounts)
}
\description{
Sums all trnascript level expression estimates to a single gene level estimate. Needs a table that maps transcript IDs to gene IDs.
}
