% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_status.R
\name{geneScores}
\alias{geneScores}
\title{score indvidual genes}
\usage{
geneScores(queryMatrix, ctt, trainNorm, grn_all, importantGenes)
}
\arguments{
\item{queryMatrix}{query ranked expression matrix}

\item{ctt}{the name of the cancer type specific network}

\item{trainNorm}{the normalization statistics from \code{\link{ccn_trainNorm}}}

\item{grn_all}{the grn construction results from \code{\link{ccn_makeGRN}}}

\item{importantGenes}{a list of gene importances}
}
\value{
a matrix with individual gene z score for each query sample, and the expression direction it should move to achieve ideal similarity with subnetwork. The order of the gene is ranked by importance from the classifier
}
\description{
score individual genes
}
