% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_status.R
\name{geneStats}
\alias{geneStats}
\title{score indvidual genes}
\usage{
geneStats(queryMatrix, ctt, trainNorm, grn_all)
}
\arguments{
\item{queryMatrix}{query ranked expression matrix}

\item{ctt}{the name of the cancer type specific network}

\item{trainNorm}{the normalization statistics from \code{\link{ccn_trainNorm}}}

\item{grn_all}{the grn construction results from \code{\link{ccn_makeGRN}}}
}
\value{
a matrix with individual gene score for each query sample
}
\description{
score individual genes
}
