% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gnrAll}
\alias{gnrAll}
\title{find genes higher in a cluster compared to all other cells}
\usage{
gnrAll(expDat, cellLabels)
}
\arguments{
\item{expDat}{a matrix of normalized gene expression taken from \code{\link(trans_prop)}}

\item{cellLabels}{a vector of all the named vector of cancer categories}
}
\value{
list of dataFrames containing pval, cval and holm for each gene in each cancer category
}
\description{
ind genes higher in a cluster compared to all other cells
}
