% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_status.R
\name{grn_status_query}
\alias{grn_status_query}
\title{Get Query GRN status}
\usage{
grn_status_query(expQuery, expTrain, stTrain, grn_return,
  classifier_return, dLevel, sidCol, classWeight = TRUE,
  exprWeight = FALSE, prune = TRUE)
}
\arguments{
\item{expQuery}{logRanked query expression matrix}

\item{expTrain}{logRanked training expression matrix}

\item{stTrain}{sample table of training expression matrix}

\item{grn_return}{the grn list that is returned from \code{\link{ccn_makeGRN}}}

\item{classifier_return}{the classifier_return list that is returned from \code{\link{broadClass_train}}}

\item{dLevel}{the name of the column with cancer types}

\item{sidCol}{the name of the column with sample IDs}

\item{classWeight}{boolean indicating whether to take the importance of the classification into status calculation}

\item{exprWeight}{boolean indicating whether to take the weight of gene expression into status calculation}

\item{prune}{boolean indicating whether to select exclusive genes for processing classification gene importance}
}
\value{
a matrix indicating the GRN status
}
\description{
Get the GRN status of query samples
}
