% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_gpa_sel.R
\name{hm_gpa_sel}
\alias{hm_gpa_sel}
\title{Heatmap of the genes and groups}
\usage{
hm_gpa_sel(expDat, genes, grps, maxPerGrp = 100, cRow = FALSE,
  cCol = FALSE, limits = c(0, 10), toScale = FALSE,
  fontsize_row = 4, reOrderCells = FALSE, ...)
}
\arguments{
\item{expDat}{normalized expression matrix from \code{\link{trans_prop}}}

\item{genes}{classification worth genes from \code{\link{findClassyGenes}}}

\item{grps}{a vector that maps sample names to a cancer category}

\item{maxPerGrp}{the number max samples per group}

\item{cRow}{TRUE if the rows should be clustered}

\item{cCol}{TRUE if the columns should be clustered}

\item{limits}{a vector of size 2 indicating the lowest and highest range for expression level}

\item{toScale}{TRUE if the data should be scaled}

\item{fontsize_row}{the font size for the row labels}
}
\value{
a heatmap of genes and their groups
}
\description{
Generate a heatmap with classification worthy genes and cancer groups
}
