% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_utils.R
\name{logRank}
\alias{logRank}
\title{logRank gene expression}
\usage{
logRank(expTrain, base = 0)
}
\arguments{
\item{expTrain}{the expression matrix}

\item{base}{the base of the log. Default is 0 which means it will just return the rank.}
}
\value{
the log10 rank of genes
}
\description{
rank the gene expression values and log10 the ranks
}
