% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptGetTop.R
\name{makePairTab}
\alias{makePairTab}
\title{Make the pair tabs}
\usage{
makePairTab(genes)
}
\arguments{
\item{genes}{a vector of all the genes in the expression matrix}
}
\description{
Generate all the combination of gene pairs
}
