% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSubClassifier.R
\name{makeSubClassifier}
\alias{makeSubClassifier}
\title{Make Classifier tailored for subclassifier}
\usage{
makeSubClassifier(expTrain, genes, groups, nRand, ntrees = 2000,
  classMatrix_rand, pairTransformedMatrix)
}
\arguments{
\item{expTrain}{transformed training data from \code{\link{query_transform}}}

\item{genes}{vector of gene pairs from \code{\link{ptGetTop}} used as predictors}

\item{groups}{named vector of cells to cancer categories}

\item{nRand}{number of randomized profiles to make}

\item{ntrees}{number of trees to build}

\item{classMatrix_rand}{the broad classification for the random category}

\item{pairTransformedMatrix}{the pairtransformed matrix of the samples}
}
\value{
Random Forest Classifier object for sub-classifier
}
\description{
Create a random forest classifier with the transformed training data from \code{\link{query_transform}}.
}
