% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_status.R
\name{meanTraining}
\alias{meanTraining}
\title{calculate the mean GRN scores across categories}
\usage{
meanTraining(grnScores, stTrain, dLevel, sidCol)
}
\arguments{
\item{grnScores}{the GRN scores calculated through ccn_score}

\item{stTrain}{the sample table used for training}

\item{dLevel}{the name of the column with all the categories}
}
\value{
an averaged GRN score matrix
}
\description{
calculate the mean GRN scores across categories
}
