% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifiedRandomize.R
\name{ModifiedRandomize}
\alias{ModifiedRandomize}
\title{Randomize the data matrix}
\usage{
ModifiedRandomize(expDat, num = 50)
}
\arguments{
\item{expDat}{the data matrix}

\item{num}{number of random profiles}
}
\value{
a randomized matrix
}
\description{
To generate random profiles by providing random permutations of the data matrix
}
