% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_transform.R
\name{pair_transform}
\alias{pair_transform}
\title{Gene to Gene comparison}
\usage{
pair_transform(expDat)
}
\arguments{
\item{expDat}{raw count expression table with cgenes}
}
\value{
a dataframe with individual gene pairs on rows and samples as columns
}
\description{
Make complete gene to gene comparison for all the classification worthy genes among each sample
}
