% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptGetTop.R
\name{parallel_quickPairs}
\alias{parallel_quickPairs}
\title{function for parallel computation of quick pairs (non-windows...)}
\usage{
parallel_quickPairs(inputPackage, topX)
}
\arguments{
\item{inputPackage}{a list of 2 things: template vectors and gene pair matrix}

\item{topX}{number of top pairs selection}
}
\value{
top scoring gene pairs
}
\description{
this function was written to compute the gene pairs in parallel for quick pairs
}
