% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_class_PRs.R
\name{plot_class_PRs}
\alias{plot_class_PRs}
\title{Plot precision recall curve per category}
\usage{
plot_class_PRs(assessed)
}
\arguments{
\item{assessed}{result of running \code{\link{ccn_classAssess}}}
}
\value{
ggplot pbject
}
\description{
Plotting the results of ccn_classAssess
}
\examples{
testAssTues<-cn_splitMakeAssess(stTrain, expTrain, ctGRNs, prop=.5)
plot_class_PRs(testAssTues$ROCs)

}
