% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneComparison.R
\name{plotGeneComparison}
\alias{plotGeneComparison}
\title{Gene expression plotting}
\usage{
plotGeneComparison(expDat, fontsize_row = 6, grps = NULL,
  cRows = FALSE, cCols = FALSE, toScale = FALSE, ...)
}
\arguments{
\item{expDat}{comparison expression matrix from \code{\link{makeGeneCompareTab}}}

\item{fontsize_row}{row font size}

\item{grps}{the groups for the heatmap}

\item{cRows}{TRUE if cluster the rows}

\item{cCols}{TRUE if cluster the columns}

\item{...}{pass parameters to pheamap}
}
\value{
a heatmap of gene expressions for the comparsion expression matrix
}
\description{
Plot gene expressions for visualization and comparison
}
