% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_utils.R
\name{processImportance}
\alias{processImportance}
\title{process classifier importance}
\usage{
processImportance(classifier, xpairs, prune = TRUE)
}
\arguments{
\item{classifier}{the Random Forest classifier}

\item{xpairs}{the xpair_list from training}

\item{prune}{blooln indicate whether to select genes exclusive to cancer category}
}
\value{
a list of genes with their importance divided by groups
}
\description{
process the importance scores of gene pairs into enriched genes
}
