% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_status.R
\name{proportionGeneExpression}
\alias{proportionGeneExpression}
\title{Proportionize Gene Expression
Calculate the proportion of expression}
\usage{
proportionGeneExpression(expTab, scaleFactor = 1000)
}
\arguments{
\item{expTab}{Gene expression matrix}

\item{scaleFactor}{Scaling factor}
}
\value{
Gene Expression proportion
}
\description{
Proportionize Gene Expression
Calculate the proportion of expression
}
