% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptGetTop.R
\name{ptGetTop}
\alias{ptGetTop}
\title{Find the best gene pairs for training (new code)}
\usage{
ptGetTop(expDat, cell_labels, topX = 50, sliceSize = 5000)
}
\arguments{
\item{expDat}{expDat}

\item{cell_labels}{named vector, value is grp, name is cell name}

\item{topX}{50}

\item{sliceSize}{the size of the slice. Default at 5e3}
}
\value{
vector of gene-pair names
}
\description{
Find the gene pairs that most distinguish a cancer group from the rest
}
