% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptGetTop.R
\name{ptGetTop_old}
\alias{ptGetTop_old}
\title{Find the best gene pairs for training}
\usage{
ptGetTop_old(expDat, cell_labels, topX = 50, sliceSize = 5000)
}
\arguments{
\item{expDat}{raw expression data}

\item{cell_labels}{named vector, value is grp, name is cell name}

\item{topX}{number of}

\item{sliceSize}{}
}
\value{
vector of gene-pair names
}
\description{
Find the gene pairs that most distinguish a cancer group from the rest
}
