% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptGetTop.R
\name{ptSmall}
\alias{ptSmall}
\title{Pair Transform on small scale}
\usage{
ptSmall(expDat, pTab)
}
\arguments{
\item{expDat}{the gene expression dataframe}

\item{pTab}{the gene pair table generated as one of the intermediate step from \code{\link{ptGetTop}}}
}
\value{
a dataframe with gene pairs as rows and samples as columns
}
\description{
Performs gene pair comparison on a smaller subset to conserve RAM
}
