<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize}
\alias{randomize}
\title{Randomize the data matrix}
\usage{
randomize(expDat, num = 50)
}
\arguments{
\item{expDat}{the data matrix}

\item{num}{number of random profiles}
}
\value{
a randomized matrix
}
\description{
To generate random profiles by providing random permutations of the data matrix
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCN.R
\name{randomize}
\alias{randomize}
\title{randomize data matrix}
\usage{
randomize(expDat, num = 50)
}
\arguments{
\item{expDat}{expDat}

\item{num}{number of profiles to return}
}
\value{
exp matrix random
}
\description{
randomize data matrix
}
>>>>>>> parent of 590b901... Documentation
