% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTab.R
\name{sc_compAlpha}
\alias{sc_compAlpha}
\title{Compute Alpha}
\usage{
sc_compAlpha(expMat, threshold = 0, pseudo = FALSE)
}
\arguments{
\item{expMat}{a matrix containing gene expression data}

\item{threshold}{a number indicating the detection threshold}

\item{pseudo}{logical indicating if it is going to be pseudo alpha or real alpha}
}
\value{
a list of alphas for the genes in expression matrix
}
\description{
Compute Alpha of expression matrix given detection threshold
}
