% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTab.R
\name{sc_compMu}
\alias{sc_compMu}
\title{compute Mu}
\usage{
sc_compMu(expMat, threshold = 0)
}
\arguments{
\item{expMat}{a matrix with gene expressions data}

\item{threshold}{a number indicating the threshold at which gene expression below does not count}
}
\value{
a list of Mus for individual gene in the expression matrix
}
\description{
compute Mu given threshold of the expression matrix
}
