% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTab.R
\name{sc_cov}
\alias{sc_cov}
\title{Compute Coefficient of Variation}
\usage{
sc_cov(vector)
}
\arguments{
\item{vector}{a vector}
}
\value{
the coefficient of variation
}
\description{
Compute the coefficient of variation of a vector
}
