% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampR_to_pattern.R
\name{sc_sampR_to_pattern}
\alias{sc_sampR_to_pattern}
\title{Create Template}
\usage{
sc_sampR_to_pattern(sampR)
}
\arguments{
\item{sampR}{a vector of the sample catgories}
}
\value{
a list of vectors indicating mapping of each cancer type in the sampR vector
}
\description{
Create a template of the cancer sample in terms of cancer category.
}
