% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPattern.R
\name{sc_testPattern}
\alias{sc_testPattern}
\title{Template Matching}
\usage{
sc_testPattern(pattern, expDat)
}
\arguments{
\item{pattern}{the template pattern created from \code{\link{sc_sampR_to_pattern}}}

\item{expDat}{normalized expression data}
}
\value{
a dataframe of genes with pval, cval and holms
}
\description{
Match the template created \code{\link{sc_sampR_to_pattern}} with expression data
}
