% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdGeneCat.R
\name{sdGeneCat}
\alias{sdGeneCat}
\title{Standard Deviation Category Gene Expressions}
\usage{
sdGeneCat(expDat, sampTab, dLevel, sampID = NULL)
}
\arguments{
\item{expDat}{the normalized expression table from \code{\link{trans_prop}}}

\item{sampTab}{the sample table}

\item{dLevel}{the name of the column that contains categories}

\item{sampID}{the name of the column that contains sample ID}
}
\value{
a matrix of the standard deviation for gene expressions
}
\description{
Finding the standard deviation the gene expression level for each category
}
