% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitCommon}
\alias{splitCommon}
\title{split a sample table into training and validation}
\usage{
splitCommon(sampTab, ncells, dLevel = "description1")
}
\arguments{
\item{sampTab}{sampTab}

\item{ncells}{ncells}

\item{dLevel}{dLevel}
}
\value{
list of train and val
}
\description{
split a sample table into training and validation
}
\examples{
stList<-splitCommon(stTrain, ncells=25, dLevel="description1")
}
