% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitCommon_proportion.R
\name{splitCommon_proportion}
\alias{splitCommon_proportion}
\title{Split training set and validation set}
\usage{
splitCommon_proportion(sampTab, proportion, dLevel)
}
\arguments{
\item{sampTab}{the sample table}

\item{proportion}{the fraction of all available data in a category that you want to use for training}

\item{dLevel}{the name of the sample table column with the classification categories}
}
\value{
a list containing training sample table and validation sample table
}
\description{
Split the sample table into training and validation by proportion
}
