% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subClass_hmClass.R
\name{subClass_hmClass}
\alias{subClass_hmClass}
\title{Detailed Complex heatmap for subclass classification}
\usage{
subClass_hmClass(subClassMat, broadClassMat, subCat, cScoreCat,
  thresHold_list = NULL)
}
\arguments{
\item{subClassMat}{classification matrix of cancer sub-category}

\item{cScoreCat}{the broad category that users want to display the classification score}

\item{thresHold_list}{list of threshold for broad class classification}

\item{broadClassMAt}{classification matrix of cancer broad category}
}
\value{
classification heatmap
}
\description{
Generates a complex heatmap for visualizing subclass classification
}
