% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subClass_predict.R
\name{subClass_predict}
\alias{subClass_predict}
\title{Predict sub class query using broad class and subclass classifier}
\usage{
subClass_predict(cnProc_broad, cnProc_sub, expDat, nrand = 2)
}
\arguments{
\item{cnProc_broad}{the cnProc of the broad classifier}

\item{expDat}{the expression data of query data}

\item{nrand}{the number of random profiles generate for evaluation process}

\item{cnProc_subclass}{the cnProc of the subclass classifier}
}
\value{
a classification score matrix
}
\description{
The function predicts the query data using the broad class and subclass classifier
}
