% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subClass_predictNew.R
\name{subClass_predictNew}
\alias{subClass_predictNew}
\title{Predict sub class query using broad class and subclass classifier (newer version)}
\usage{
subClass_predictNew(cnProc_broad, cnProc_sub, expDat, nrand = 2,
  weight_broadClass = 1)
}
\arguments{
\item{cnProc_broad}{the cnProc of the broad classifier}

\item{expDat}{the expression data of query data}

\item{nrand}{the number of random profiles generate for evaluation process}

\item{weight_broadClass}{the weight that will be put on the broadclass classification}

\item{cnProc_subclass}{the cnProc of the subclass classifier}
}
\value{
a classification score matrix
}
\description{
The function predicts the query data using the broad class and subclass classifier
}
