% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subClass_train.R
\name{subClass_train}
\alias{subClass_train}
\title{Sub-Class Training (new and improved version) with merged rand category}
\usage{
subClass_train(cnProc_broad, stTrain, expTrain, colName_broadCat,
  colName_subClass, name_broadCat, colName_samp = "row.names",
  nTopGenes = 20, nTopGenePairs = 50, nRand = 40, nTrees = 1000,
  weightedDown_total = 5e+05, weightedDown_dThresh = 0.25,
  transprop_xFact = 1e+05, weight_broadClass = 1, randomCat = NULL)
}
\arguments{
\item{cnProc_broad}{the broad cnProc from \code{\link{broadClass_train}}}

\item{stTrain}{a dataframe that matches the samples with broad category and sub-class}

\item{expTrain}{the expression matrix}

\item{colName_broadCat}{the name of the column in sample table that contains broad categories}

\item{colName_subClass}{the name of the column in sample table that contains sub class}

\item{name_broadCat}{the name of the broad class in which the subclasses are}

\item{colName_samp}{the name of the column that contains sample names}

\item{nTopGenes}{the number of classification genes per category}

\item{nTopGenePairs}{the number of top gene pairs per category}

\item{nRand}{number of random profiles generate for training}

\item{nTrees}{number of trees for random forest classifier}

\item{weightedDown_dThresh}{the threshold at which anything lower than that is 0 for weighted_down function}

\item{transprop_xFact}{scaling factor for transprop}

\item{weight_broadClass}{the weight on the result of the broad classification as features for subclassifier}

\item{randomCat}{the name of the category that users want to sample from to make rand}

\item{weightDown_total}{numeric post transformation sum of read counts for weighted_down function}
}
\value{
a list containing normalized expression data, classification gene list, cnProc
}
\description{
Tranining sub-class classifier
}
