% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subClassQuery_transform.R
\name{subClassQuery_transform}
\alias{subClassQuery_transform}
\title{Sub-class Query Transform}
\usage{
subClassQuery_transform(expDat, cgenes, xpairs, classMatrix)
}
\arguments{
\item{expDat}{query expression data}

\item{cgenes}{classification genes}

\item{xpairs}{classification gene pairs}

\item{classMatrix}{classification matrix of the query expression data from broad classifier}
}
\value{
transformed sub-class query matrix
}
\description{
Transform the sub-class query expression data
}
