% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convertLongTab.R
\name{utils_convertLongTab}
\alias{utils_convertLongTab}
\title{Convert Long Expression Table}
\usage{
utils_convertLongTab(longDf, SampCol, geneCol, geneExpCol,
  geneNaOmit = TRUE)
}
\arguments{
\item{longDf}{a dataframe of gene expression/count in long table format}

\item{SampCol}{the name of the column containing sample names}

\item{geneCol}{the name of the column containing gene names}

\item{geneNaOmit}{TRUE if you want to omit the genes with NA}

\item{geneExpcol}{the name of the column containing gene counts or expression}
}
\value{
a list containing the gene expression dataframe in wide table format and a dataframe of omitted genes if geneNaOmit is TRUE
Otherwise, just returns gene Expression matrix in wide table
}
\description{
Convert long gene expression/count table into wide dataframe to proceed to the next step
of training classifier or validating samples.
}
