% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_matchSampTab.R
\name{utils_matchSampTab}
\alias{utils_matchSampTab}
\title{Match Sample Table}
\usage{
utils_matchSampTab(expDat, sampTab, sampID, sampDesc)
}
\arguments{
\item{expDat}{the expression matrix}

\item{sampTab}{the sample table}

\item{sampID}{the name of the column in sampTab that contains the sample names}

\item{sampDesc}{the name of the column in sampTab that contains the sample groups/description}
}
\value{
a list of ordered ordered expression matrix, sample table. If there are unmatched expression matrix or sample tabble
the list will also contain unmatched expression matrix and unmatched sample table.
}
\description{
To match the ordering of the sample names in expression matrix with the sample table.
}
