% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_status.R
\name{utils_reduceMat}
\alias{utils_reduceMat}
\title{Reduce data matrix}
\usage{
utils_reduceMat(datFrame, valCol, cName = "ann")
}
\arguments{
\item{datFrame}{the dataframe}

\item{valCol}{the column with scores}

\item{cName}{column with groups}
}
\value{
df of grp_name, mean, sd
}
\description{
reduce the data.matrix values by averaging and getting st dvs
}
