% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_status.R
\name{zscore}
\alias{zscore}
\title{Z scores
Figure out Z score given mean and standard deviation}
\usage{
zscore(x, meanVal, sdVal)
}
\arguments{
\item{x}{query score}

\item{meanVal}{mean of the distribution}
}
\description{
Z scores
Figure out Z score given mean and standard deviation
}
