% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{addRandToSampTab}
\alias{addRandToSampTab}
\title{adds 'random' samples to a sample table to help with plotting}
\usage{
addRandToSampTab(classRes, sampTab, desc, id = "cell_name")
}
\arguments{
\item{classRes}{classRes}

\item{sampTab}{sampTab}

\item{desc}{desc}

\item{id}{id}
}
\value{
sample table
}
\description{
adds 'random' samples to a sample table to help with plotting
}
\examples{
stNew<-addRandToSampTab(classRes, stOld, "description1", "sample_name")

}
