% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{addRandToSampTab}
\alias{addRandToSampTab}
\title{add "random" profiles to the meta sample table to assist with plotting}
\usage{
addRandToSampTab(classRes, sampTab, desc, id = "cell_name")
}
\arguments{
\item{classRes}{the classification matrix generated from \code{\link{rf_classPredict}}}

\item{sampTab}{the meta sample table}

\item{desc}{the "column name" of the column containing the sample descriptions in the meta sample table}

\item{id}{the column name of the column containing the sample names in the meta sample table}
}
\value{
meta sample table containing the random profiles
}
\description{
This will add the random profiles generated from \code{\link{rf_classPredict}} to the meta sample table
}
\examples{
stValRand<-addRandToSampTab(classRes_val, stVal, "description2", "sample_name")
}
