% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ccn_hmClass}
\alias{ccn_hmClass}
\title{Heatmap of the classification result}
\usage{
ccn_hmClass(classMat, grps = NULL, isBig = FALSE, cRow = FALSE,
  cCol = FALSE, fontsize_row = 4, scale = FALSE)
}
\arguments{
\item{classMat}{classification matrix generated from \code{\link{rf_classPredict}}}

\item{isBig}{is this a big heatmap? TRUE or FALSE}

\item{cluster_cols}{cluster_cols}
}
\value{
nothing
}
\description{
This function generates a heatmap of the classification result for visualization
}
\examples{
ccn_HmClass(cnRes, isBig=TRUE)

}
