% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ccn_hmClass}
\alias{ccn_hmClass}
\title{heatmap of the classification result}
\usage{
ccn_hmClass(classMat, grps = NULL, isBig = FALSE, cRow = FALSE,
  cCol = FALSE, fontsize_row = 4, scale = FALSE)
}
\arguments{
\item{classMat}{classMat}

\item{isBig}{is this a big heatmap? TRUE or FALSE}

\item{cluster_cols}{cluster_cols}
}
\value{
nothing
}
\description{
Heatmap of the classification result.
}
\examples{
ccn_HmClass(cnRes, isBig=TRUE)

}
