% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn_hmClass.R
\name{ccn_hmClass}
\alias{ccn_hmClass}
\title{Heatmap of the classification result}
\usage{
ccn_hmClass(classMat, grps = NULL, isBig = FALSE, cRow = FALSE,
  cCol = FALSE, fontsize_row = 4, fontsize_col = 4, main = NA,
  scale = FALSE)
}
\arguments{
\item{classMat}{classification matrix generated from \code{\link{rf_classPredict}}}

\item{grps}{a vector that maps samples to a group}

\item{isBig}{TRUE if this is a big heatmap}

\item{cRow}{TRUE if rows should be clustered}

\item{cCol}{TRUE if columns should be clustered}

\item{fontsize_row}{the font size of the row}

\item{fontsize_col}{the font size of the columns}

\item{main}{the title of the heatmap}

\item{scale}{FALSE if the highest value is 1 and the lowest is 0}
}
\value{
classification heatmap
}
\description{
This function generates a heatmap of the classification result for visualization
}
\examples{
ccn_HmClass(cnRes, isBig=TRUE)

}
