% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn_status.R
\name{ccn_rawScore}
\alias{ccn_rawScore}
\title{computes the raw z score for a gene as xmax-abs(zscore).}
\usage{
ccn_rawScore(vect, mmean, ssd, xmax = 1000, reg_type)
}
\arguments{
\item{vect}{a vector of gene expression values for multiple samples}

\item{mmean}{mean value in training data}

\item{ssd}{standard deviation in training data}

\item{reg_type}{the type of regulation (up or down) for the specific subnetwork}
}
\value{
transformed (but not normalized) GRN score
}
\description{
better values are higher.
}
