% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn_classAssess.R
\name{cn_eval}
\alias{cn_eval}
\title{determine performance of classification at given threshold}
\usage{
cn_eval(vect, sampTab, dLevel, classification, threshs = seq(0, 1, by =
  0.05), dLevelSID = "sample_id")
}
\arguments{
\item{vect}{vector of values}

\item{sampTab}{sample table}

\item{dLevel}{colname}

\item{classification}{actual classification}

\item{dLevelSID}{column to indicate sample id}

\item{thresh}{threshold above which to make a call}
}
\value{
return a data frame of the number of TP, FN, FP, and TN
}
\description{
determine performance of classification at given threshold
}
