% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn_hmClass.R
\name{cn_HmClass}
\alias{cn_HmClass}
\title{Simple heatmap of the classification result}
\usage{
cn_HmClass(classRes, isBig = FALSE, cRow = FALSE, cCol = FALSE,
  fontsize_row = 4, scale = FALSE)
}
\arguments{
\item{isBig}{TRUE if this is a big heatmap}

\item{cRow}{TRUE if rows should be clustered}

\item{cCol}{TRUE if columns should be clustered}

\item{fontsize_row}{the font size}

\item{scale}{FALSE if the highest value is 1 and the lowest is 0}

\item{classMat}{classification matrix generated from \code{\link{rf_classPredict}}}
}
\value{
classification heatmap
}
\description{
Heatmap of the classification result that does not have an explicit group.
}
