% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_transform.R
\name{query_transform}
\alias{query_transform}
\title{Make complete gene-to-gene comparison}
\usage{
query_transform(expDat, genePairs)
}
\arguments{
\item{expDat}{the expression matrix}

\item{genePairs}{a vector with gene pairs}
}
\value{
A matrix of the gene to gene comparison results. 1 indicates the first gene has higher expression than the second gene in the gene pair. 0 indicates the
second gene has higher expression than the first gene in the gene pair.
}
\description{
Performs gene to gene comparison of each gene pairs in the expression table.
}
