% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCN.R
\name{query_transform}
\alias{query_transform}
\title{Make complete gene-to-gene comparison}
\usage{
query_transform(expDat, genePairs)
}
\arguments{
\item{expDat}{the expression matrix}

\item{genePairs}{a vector with gene pairs}
}
\value{
A matrix indicating whether the first gene in the gene pair has a greater expression than the second gene in the gene pair
}
\description{
This function will compare the genes in genePairs vector sample by sample in the expression matrix.
}
