% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_classPredict.R
\name{rf_classPredict}
\alias{rf_classPredict}
\title{Classify Query Samples}
\usage{
rf_classPredict(rfObj, expQuery, numRand = 50)
}
\arguments{
\item{rfObj}{random forest object created from \code{\link{makeClassifier}}}

\item{expQuery}{transformed query data created from \code{\link{query_transform}}}

\item{numRand}{number of random profiles to add into the classification}
}
\value{
a classification matrix of the query data
}
\description{
Classify the transformed query data using the random forest classifier.
}
