% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_classPredict.R
\name{rf_classPredict}
\alias{rf_classPredict}
\title{Classify the samples}
\usage{
rf_classPredict(rfObj, expQuery, numRand = 50)
}
\arguments{
\item{rfObj}{random forest object created from \code{\link{makeClassifier}}}

\item{expQuery}{transformed query data created from \code{\link{query_transform}}}

\item{numRand}{number of random profiles wish to create}
}
\value{
a classification matrix of the query data
}
\description{
This functions will classify the query samples using the random forest object created from \code{\link{makeClassifier}}
}
