% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitCommon.R
\name{splitCommon}
\alias{splitCommon}
\title{Split Sample Table}
\usage{
splitCommon(sampTab, ncells, dLevel = "description1")
}
\arguments{
\item{sampTab}{sample table (DataFrame)}

\item{ncells}{number of samples for training in each category (Integer)}

\item{dLevel}{the column name with the classification categories (String)}
}
\value{
a list containing training sample table and validation sample table
}
\description{
Split a sample table into training set and validation set.
}
\examples{
stList<-splitCommon(stTrain, ncells=25, dLevel="description1")
}
