% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitCommon.R
\name{splitCommon}
\alias{splitCommon}
\title{Split Sample Table Randomly}
\usage{
splitCommon(sampTab, ncells, dLevel = "description1")
}
\arguments{
\item{sampTab}{sample table}

\item{ncells}{number of samples for training in each category}

\item{dLevel}{the column name with the classification categories}
}
\value{
a list containing training sample table and validation sample table
}
\description{
Split the sample table into training set and validation set in random fashion.
}
