% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_prop.R
\name{trans_prop}
\alias{trans_prop}
\title{Log scaled downsampled expression}
\usage{
trans_prop(expDat, xFact = 1e+05)
}
\arguments{
\item{expDat}{a matrix of weighted down expression data}

\item{xFact}{a number representing scaling factor}
}
\value{
a matrix of downsampled read mapped to genes/transcripts
}
\description{
Divide each column by sum of that column then scale to xFact and log it for more normalization.
}
