% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convertToGeneSymbols.R
\name{utils_convertToGeneSymbols}
\alias{utils_convertToGeneSymbols}
\title{Gene Symbol Conversion}
\usage{
utils_convertToGeneSymbols(expTab, typeENST = FALSE, typeENSG = FALSE,
  typeMusGene = FALSE)
}
\arguments{
\item{expTab}{the matrix of experssion data with sample names as column names and gene names as row names}

\item{typeENST}{TRUE if the gene identifier is in Ensembl Transcript ID}

\item{typeENSG}{TRUE if the gene identifier is in Ensembl Gene ID}

\item{typeMusGene}{TRUE if the gene identifier is in mouse gene symbols}
}
\value{
a converted expression matrix
}
\description{
The function can convert Ensembl Gene ID, Ensembl Transcript ID or mouse gene symbol into human gene symbol
}
