% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_loadObject.R
\name{utils_loadObject}
\alias{utils_loadObject}
\title{Load an R object}
\usage{
utils_loadObject(fname)
}
\arguments{
\item{fname}{The file name of the R object}
}
\value{
A R object
}
\description{
Load an R object
}
